feat: Initializing the proyect for the actual objective of the repo
